refactor: split script into ES modules
