+++ index.html
@@ -173,6 +173,6 @@
     <script src="particles-configuration.js"></script>
     <script src="language.js"></script>
     <script src="settings.js"></script>
-    <script src="script.js"></script>
+    <script type="module" src="ui.js"></script>
   </body>
 </html>
